Refactor production.py (#187)
